@@ -1,10 +1,12 @@
 plugins {
     alias(libs.plugins.android.application)
+    id("com.google.gms.google-services")
 }
 
 android {
     namespace = "com.example.petpal"
-    compileSdk = 34
+    compileSdk = 35
+
 
     defaultConfig {
         applicationId = "com.example.petpal"
@@ -28,13 +30,31 @@ android {
     compileOptions {
         sourceCompatibility = JavaVersion.VERSION_1_8
         targetCompatibility = JavaVersion.VERSION_1_8
+        encoding = "UTF-8"
     }
+
 }
 
 dependencies {
+    implementation ("androidx.work:work-runtime-ktx:2.8.1")
+    implementation ("org.eclipse.paho:org.eclipse.paho.client.mqttv3:1.2.5")
     implementation ("androidx.room:room-runtime:2.3.0")
-    annotationProcessor ("androidx.room:room-compiler:2.3.0")
+    implementation(libs.firebase.firestore)
+    implementation(libs.places)
+    annotationProcessor("androidx.room:room-compiler:2.3.0")
     implementation("androidx.recyclerview:recyclerview:1.2.1")
+    implementation("com.google.android.gms:play-services-maps:19.0.0")
+    implementation("com.google.android.gms:play-services-location:21.3.0")
+    implementation(platform("com.google.firebase:firebase-bom:33.6.0"))
+    implementation("com.google.firebase:firebase-database:21.0.0")
+    implementation("com.google.firebase:firebase-analytics:22.1.2")
+    implementation("com.google.firebase:firebase-bom:33.6.0")
+    implementation("com.google.firebase:firebase-auth:23.1.0")
+    implementation ("com.google.android.gms:play-services-places:17.1.0")
+    implementation ("com.github.bumptech.glide:glide:4.12.0")
+    implementation ("com.google.code.gson:gson:2.8.8")
+    implementation(libs.core.ktx)
+    implementation(libs.firebase.database)
     implementation(libs.appcompat)
     implementation(libs.material)
     implementation(libs.activity)
