@@ -1,6 +1,7 @@
 package com.example.petpal;
 
 import android.content.Intent;
+import android.content.SharedPreferences;
 import android.os.Bundle;
 import android.text.TextUtils;
 import android.view.View;
@@ -12,30 +13,37 @@ import android.widget.Toast;
 
 import androidx.appcompat.app.AppCompatActivity;
 
+import com.google.firebase.firestore.FirebaseFirestore;
+import com.google.firebase.firestore.QuerySnapshot;
+import com.google.firebase.firestore.DocumentSnapshot;
+
 public class Login extends AppCompatActivity {
 
     private EditText emailEditText, passwordEditText;
     private Button loginButton;
     private ProgressBar progressBar;
     private TextView registerTextView;
-    private DatabaseHelper databaseHelper;
+    private FirebaseFirestore db;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_login);
 
-        // Inicializar el DatabaseHelper
-        databaseHelper = new DatabaseHelper(this);
+        db = FirebaseFirestore.getInstance();
 
-        // Referencias a los campos de entrada y el botón
         emailEditText = findViewById(R.id.email_input);
         passwordEditText = findViewById(R.id.password_input);
         loginButton = findViewById(R.id.login_button);
         progressBar = findViewById(R.id.progress_bar);
         registerTextView = findViewById(R.id.register_link);
 
-        // Acción al hacer clic en el botón de iniciar sesión
+        registerTextView.setOnClickListener(v -> {
+            // Redirige a la actividad de Registro
+            Intent intent = new Intent(Login.this, Registro.class);
+            startActivity(intent);
+        });
+
         loginButton.setOnClickListener(v -> {
             String email = emailEditText.getText().toString().trim();
             String password = passwordEditText.getText().toString().trim();
@@ -50,60 +58,42 @@ public class Login extends AppCompatActivity {
                 return;
             }
 
-            // Verificación de credenciales de administrador
-            if (email.equals("admin@gmail.com") && password.equals("123")) {
-                // Mostrar el ProgressBar
-                progressBar.setVisibility(View.VISIBLE);
+            progressBar.setVisibility(View.VISIBLE);
 
-                // Ejecutar en un hilo separado para simular un tiempo de espera
-                new Thread(() -> {
-                    try {
-                        Thread.sleep(3000); // Espera de 3 segundos
-
-                        runOnUiThread(() -> {
-                            progressBar.setVisibility(View.GONE);  // Ocultar el ProgressBar
-                            Toast.makeText(Login.this, "Inicio de sesión exitoso", Toast.LENGTH_SHORT).show();
-
-                            // Redirigir a la actividad UserListActivity si es el administrador
-                            Intent intent = new Intent(Login.this, activity_user_list.class);
-                            startActivity(intent);
-                            finish();
-                        });
-                    } catch (InterruptedException e) {
-                        e.printStackTrace();
-                    }
-                }).start();
-            } else {
-                // Lógica de inicio de sesión normal para otros usuarios
-                progressBar.setVisibility(View.VISIBLE);
-
-                // Verifica las credenciales en la base de datos
-                new Thread(() -> {
-                    boolean isValidUser = databaseHelper.verifyUserCredentials(email, password);
-
-                    runOnUiThread(() -> {
+            db.collection("usuarios")
+                    .whereEqualTo("email", email)
+                    .whereEqualTo("password", password)
+                    .get()
+                    .addOnCompleteListener(task -> {
                         progressBar.setVisibility(View.GONE);
-                        if (isValidUser) {
-                            Toast.makeText(Login.this, "Inicio de sesión exitoso", Toast.LENGTH_SHORT).show();
 
-                            // Redirigir al MenuPrincipal para otros usuarios
+                        if (task.isSuccessful() && !task.getResult().isEmpty()) {
+                            QuerySnapshot querySnapshot = task.getResult();
+                            DocumentSnapshot documentSnapshot = querySnapshot.getDocuments().get(0);
+
+                            String petName = documentSnapshot.getString("petName");
+                            String ownerName = documentSnapshot.getString("ownerName");
+                            String ownerEmail = documentSnapshot.getString("email");
+                            String profileImageUrl = documentSnapshot.getString("profileImage");
+
+                            // Guardar los datos del usuario, incluida la imagen de perfil, en SharedPreferences
+                            SharedPreferences sharedPreferences = getSharedPreferences("UserPrefs", MODE_PRIVATE);
+                            SharedPreferences.Editor editor = sharedPreferences.edit();
+                            editor.putString("petName", petName);
+                            editor.putString("ownerName", ownerName);
+                            editor.putString("ownerEmail", ownerEmail);
+                            editor.putString("profileImage", profileImageUrl);  // Guardar la URL de la imagen
+                            editor.apply();
+
+                            Toast.makeText(Login.this, "Bienvenido " + ownerName, Toast.LENGTH_SHORT).show();
+
                             Intent intent = new Intent(Login.this, MenuPrincipal.class);
                             startActivity(intent);
                             finish();
                         } else {
-                            // Mensaje de error si las credenciales son incorrectas
                             Toast.makeText(Login.this, "Credenciales incorrectas. Intenta nuevamente.", Toast.LENGTH_SHORT).show();
                         }
                     });
-                }).start();
-            }
-        });
-
-        // Acción al hacer clic en el TextView para registrarse
-        registerTextView.setOnClickListener(v -> {
-            // Ir a la actividad de registro
-            Intent intent = new Intent(Login.this, Registro.class);
-            startActivity(intent);
         });
     }
 }
